requirements.txt
Django>3.1,<3.2
requests==2.25.0
djangorestframework>3.12,<3.13
drf-yasg==1.20.0
django-admin-shortcuts==2.0.0
